+++ cli/src/commands/operation/log.rs
@@ -49,14 +49,6 @@ pub struct OperationLogArgs {
     /// Limit number of operations to show
     #[arg(long, short = 'n')]
     limit: Option<usize>,
-    // TODO: Delete `-l` alias in jj 0.25+
-    #[arg(
-        short = 'l',
-        hide = true,
-        conflicts_with = "limit",
-        value_name = "LIMIT"
-    )]
-    deprecated_limit: Option<usize>,
     /// Don't show the graph, show a flat list of operations
     #[arg(long)]
     no_graph: bool,
@@ -197,13 +189,7 @@ fn do_op_log(
     ui.request_pager();
     let mut formatter = ui.stdout_formatter();
     let formatter = formatter.as_mut();
-    if args.deprecated_limit.is_some() {
-        writeln!(
-            ui.warning_default(),
-            "The -l shorthand is deprecated, use -n instead."
-        )?;
-    }
-    let limit = args.limit.or(args.deprecated_limit).unwrap_or(usize::MAX);
+    let limit = args.limit.unwrap_or(usize::MAX);
     let iter = op_walk::walk_ancestors(slice::from_ref(current_op)).take(limit);
     if !args.no_graph {
         let mut raw_output = formatter.raw()?;